+++ src/main/java/mods/belgabor/amtweaker/AMTweaker.java
@@ -1,35 +1,36 @@
 package mods.belgabor.amtweaker;
 
+import cpw.mods.fml.common.Loader;
 import cpw.mods.fml.common.event.FMLServerStartingEvent;
-import minetweaker.MineTweakerAPI;
 import minetweaker.MineTweakerImplementationAPI;
 import minetweaker.util.IEventHandler;
 import mods.belgabor.amtweaker.mods.amt.AMT;
-import mods.belgabor.amtweaker.mods.amt.loggers.CommandLogger;
+import mods.belgabor.amtweaker.mods.amt.loggers.AMTCommandLogger;
 import mods.belgabor.amtweaker.util.TweakerPlugin;
 import cpw.mods.fml.common.Mod;
 import cpw.mods.fml.common.Mod.EventHandler;
 import cpw.mods.fml.common.event.FMLInitializationEvent;
 
-@Mod(modid = AMTweaker.MODID, version = AMTweaker.VERSION, name = "AMTweaker", dependencies = "required-after:MineTweaker3;required-after:DCsAppleMilk;after:ModTweaker;after:AWWayofTime")
+@Mod(modid = AMTweaker.MODID, version = AMTweaker.VERSION, name = "AMTweaker", dependencies = "required-after:MineTweaker3;required-after:DCsAppleMilk;after:ModTweaker;after:AWWayofTime;after:DCsEcoMT")
 public class AMTweaker implements IEventHandler<MineTweakerImplementationAPI.ReloadEvent>
 {
     public static final String MODID = "AMTweaker";
-    public static final String VERSION = "0.31";
+    public static final String VERSION = "0.6";
     
     @EventHandler
     public void init(FMLInitializationEvent event)
     {
-        TweakerPlugin.register("DCsAppleMilk", AMT.class);
+        TweakerPlugin.register(AMT.MODID, AMT.class);
     }
 
     @EventHandler
     public void onServerStart(FMLServerStartingEvent event) {
-        CommandLogger.register();
+        AMTCommandLogger.register();
         MineTweakerImplementationAPI.onReloadEvent(this);
     }
 
     public void handle(MineTweakerImplementationAPI.ReloadEvent event) {
-        CommandLogger.register();
+        if (Loader.isModLoaded(AMT.MODID))
+            AMTCommandLogger.register();
     }
 }