+++ src/main/java/mods/belgabor/amtweaker/mods/amt/handlers/Processor.java
@@ -5,7 +5,6 @@ import minetweaker.api.item.IIngredient;
 import minetweaker.api.item.IItemStack;
 import mods.belgabor.amtweaker.mods.amt.util.AMTListAddition;
 import mods.belgabor.amtweaker.mods.amt.util.AMTRecipeWrapper;
-import mods.belgabor.amtweaker.util.BaseListRemoval;
 import mods.belgabor.amtweaker.util.BaseListWildcardRemoval;
 import mods.defeatedcrow.api.recipe.IProcessorRecipe;
 import mods.defeatedcrow.api.recipe.RecipeRegisterManager;
@@ -25,26 +24,31 @@ import static mods.belgabor.amtweaker.helpers.StackHelper.areEqualNull;
 public class Processor {
     // Adding a new processor recipe
     @ZenMethod
+    public static void addRecipe(IItemStack output, IItemStack secondary, IIngredient[] inputs, boolean isFoodRecipe, float secondaryChance, boolean forceReturnContainer, int tier) {
+        doAddRecipe(output, secondary, inputs, isFoodRecipe, secondaryChance, forceReturnContainer, tier);
+    }
+
+    @ZenMethod
     public static void addRecipe(IItemStack output, IItemStack secondary, IIngredient[] inputs, boolean isFoodRecipe, float secondaryChance, boolean forceReturnContainer) {
-        doAddRecipe(output, secondary, inputs, isFoodRecipe, secondaryChance, forceReturnContainer);
+        doAddRecipe(output, secondary, inputs, isFoodRecipe, secondaryChance, forceReturnContainer, 1);
     }
 
     @ZenMethod
     public static void addRecipe(IItemStack output, IItemStack secondary, IIngredient[] inputs, boolean isFoodRecipe, float secondaryChance) {
-        doAddRecipe(output, secondary, inputs, isFoodRecipe, secondaryChance, false);
+        doAddRecipe(output, secondary, inputs, isFoodRecipe, secondaryChance, false, 1);
     }
 
     @ZenMethod
     public static void addRecipe(IItemStack output, IItemStack secondary, IIngredient[] inputs, boolean isFoodRecipe) {
-        doAddRecipe(output, secondary, inputs, isFoodRecipe, null, false);
+        doAddRecipe(output, secondary, inputs, isFoodRecipe, null, false, 1);
     }
 
     @ZenMethod
     public static void addRecipe(IItemStack output, IIngredient[] inputs, boolean isFoodRecipe) {
-        doAddRecipe(output, null, inputs, isFoodRecipe, null, false);
+        doAddRecipe(output, null, inputs, isFoodRecipe, null, false, 1);
     }
 
-    private static void doAddRecipe(IItemStack output, IItemStack secondary, IIngredient[] inputs, boolean isFoodRecipe, Float secondaryChance, boolean forceReturnContainer) {
+    private static void doAddRecipe(IItemStack output, IItemStack secondary, IIngredient[] inputs, boolean isFoodRecipe, Float secondaryChance, boolean forceReturnContainer, int tier) {
         if (inputs == null) {
             MineTweakerAPI.getLogger().logError("Processor: Input set must not be null!");
             return;
@@ -57,29 +61,31 @@ public class Processor {
             MineTweakerAPI.getLogger().logError("Processor: Primary and secondary output must not both be null!");
             return;
         }
-        MineTweakerAPI.apply(new Add(new ProcessorRecipeWrapper(output, secondary, inputs, isFoodRecipe, secondaryChance, forceReturnContainer)));
+        MineTweakerAPI.apply(new Add(new ProcessorRecipeWrapper(output, secondary, inputs, isFoodRecipe, secondaryChance, forceReturnContainer, tier)));
     }
 
     private static class ProcessorRecipeWrapper extends AMTRecipeWrapper {
-        private ItemStack output;
-        private ItemStack secondary;
-        private Object[] inputs;
-        private boolean isFoodRecipe;
-        private Float secondaryChance;
-        private boolean forceReturnContainer;
+        private final ItemStack output;
+        private final ItemStack secondary;
+        private final Object[] inputs;
+        private final boolean isFoodRecipe;
+        private final Float secondaryChance;
+        private final boolean forceReturnContainer;
+        private final int tier;
 
-        public ProcessorRecipeWrapper(IItemStack output, IItemStack secondary, IIngredient[] inputs, boolean isFoodRecipe, Float secondaryChance, boolean forceReturnContainer) {
+        public ProcessorRecipeWrapper(IItemStack output, IItemStack secondary, IIngredient[] inputs, boolean isFoodRecipe, Float secondaryChance, boolean forceReturnContainer, int tier) {
             this.output = toStack(output);
             this.secondary = toStack(secondary);
             this.inputs = toObjects(inputs, true);
             this.isFoodRecipe = isFoodRecipe;
             this.secondaryChance = secondaryChance;
             this.forceReturnContainer = forceReturnContainer;
+            this.tier = tier;
         }
 
         @Override
         public void register() {
-            RecipeRegisterManager.processorRecipe.addRecipe(output, isFoodRecipe, forceReturnContainer, secondary, (secondaryChance == null)?0:secondaryChance.floatValue(), inputs);
+            RecipeRegisterManager.processorRecipe.addRecipe(output, isFoodRecipe, tier, forceReturnContainer, secondary, (secondaryChance == null)?0:secondaryChance, inputs);
         }
 
         @Override