+++ src/main/java/mods/belgabor/amtweaker/mods/amt/util/BlockAddition.java
@@ -4,6 +4,7 @@ package mods.belgabor.amtweaker.mods.amt.util;
 import minetweaker.IUndoableAction;
 import minetweaker.MineTweakerAPI;
 import minetweaker.api.item.IItemStack;
+import mods.defeatedcrow.api.recipe.ICookingHeatSource;
 import net.minecraft.block.Block;
 import net.minecraft.item.ItemStack;
 import net.minecraftforge.oredict.OreDictionary;
@@ -11,16 +12,15 @@ import net.minecraftforge.oredict.OreDictionary;
 import java.util.List;
 
 import static mods.belgabor.amtweaker.helpers.InputHelper.toStack;
-import static mods.belgabor.amtweaker.helpers.StackHelper.areEqual;
 
 public abstract class BlockAddition implements IUndoableAction {
     protected final String description;
     protected final Block block;
     protected final int meta;
     protected final ItemStack item;
-    protected final List<ItemStack> list;
+    protected final List<? extends ICookingHeatSource> list;
 
-    public BlockAddition(String description, List<ItemStack> list, IItemStack stack) {
+    public BlockAddition(String description, List<? extends ICookingHeatSource> list, IItemStack stack) {
         this.description = description;
         this.item = toStack(stack);
         this.block = Block.getBlockFromItem(this.item.getItem());
@@ -44,8 +44,8 @@ public abstract class BlockAddition implements IUndoableAction {
     @Override
     public void undo() {
         boolean found = false;
-        for (ItemStack i: list) {
-            if (areEqual(i, item)) {
+        for (ICookingHeatSource i: list) {
+            if ((meta == i.getMetadata()) && (block.equals(i.getBlock()))) {
                 found = true;
                 list.remove(i);
                 break;